pyproject.toml
[project]
name = "html-form"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.3",
    "joblib>=1.5.3",
    "mysql-connector-python>=9.6.0",
    "numpy>=2.4.2",
    "pandas>=3.0.1",
    "scikit-learn>=1.8.0",
]
